Refactor GurpsActorV2 tests: break out _preUpdate and moveItem tests into their own test files.
